Change the uniqueness check
'Take another playbook Move', and 'Take a move from another playbook' are
duplicated and identical in each playbook's improvements. In order to
seed these idempotently, we need to make them unique using an
attribute not used by the improvement. We'll hide 0/negative stat limits
on the index, and use them to make the seeds unique.
* Hide unless stat_limit is positive
* Change Seeds to use stat_limit for uniqueness instead of rating